+++ .env.example
@@ -1,0 +1,12 @@
+LANGFUSE_PUBLIC_KEY="REPLACE_WITH_YOUR_PUBLIC_KEY
+LANGFUSE_SECRET_KEY="REPLACE_WITH_YOUR_SECRET_KEY"
+LANGFUSE_HOST="http://localhost:3000"
+HTTP_PORT=5000
+LLM_MODEL="mistral"
+EMBED_MODEL="BAAI/bge-small-en-v1.5"
+CHUNK_SIZE=1000
+CHUNK_OVERLAP=150
+REQUEST_TIMEOUT=60
+DOCUMENTS_DIR="../docs"
+COLLECTION_NAME="test_collection"
+PROMPT_TEMPLATE="YOUR_PROMPT_TEMPLATE_NAME"
