+++ .gitignore
@@ -9,3 +9,4 @@ venv/
 .idea/
 Pipfile.lock
 docs/*
+.env
