+++ Pipfile
@@ -15,6 +15,7 @@ llama-index-embeddings-huggingface = "==0.2.0"
 llama-index-llms-ollama = "==0.1.2"
 llama-index-vector-stores-chroma = "==0.1.6"
 langfuse = "*"
+python-dotenv = "==1.0.1"
 
 [dev-packages]
