+++ src/prompts.py
@@ -1,0 +1,18 @@
+from llama_index.core import PromptTemplate
+from langfuse import Langfuse
+from config import PROMPT_TEMPLATE
+from langfuse_integration import is_langfuse_enabled
+
+
+def get_template():
+    if is_langfuse_enabled() is True:
+        langfuse = Langfuse()
+        try:
+            prompt = langfuse.get_prompt(PROMPT_TEMPLATE)
+        except Exception as e:
+            print("Error in getting prompt template - ", e)
+            return None
+        template = prompt.compile()
+        qa_template = PromptTemplate(template)
+        return qa_template
+    return None
